requirements-polymarket.txt
httpx>=0.25.0
pydantic>=2.5.0
pandas>=2.0.0
python-dotenv>=1.0.0
chromadb>=0.4.0
langchain>=0.1.0
langchain-openai>=0.0.5
newsapi-python>=0.2.7
